+++ src/game/units/wizard/talents/constants.rs
@@ -29,8 +29,8 @@ pub(crate) fn tier_thresholds(spell: Spell) -> [u32; 3] {
         // Support spells
         Spell::GuardianCircle => [20, 70, 200],
         Spell::Haste => [30, 200, 500],
-        Spell::Teleport => [1, 2, 3],
-        Spell::RaiseTheDead => [15, 50, 150],
+        Spell::Teleport => [20, 100, 500],
+        Spell::RaiseTheDead => [1, 2, 3],
         Spell::HealingPlume => [30, 100, 300],
         Spell::FogCloud => [20, 70, 200],
         Spell::BerserkerRage => [15, 50, 150],